+++ core/src/main/java/de/bitbrain/braingdx/input/InputManager.java
@@ -2,8 +2,6 @@ package de.bitbrain.braingdx.input;
 
 import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.controllers.ControllerListener;
-import de.bitbrain.braingdx.behavior.movement.Movement;
-import de.bitbrain.braingdx.behavior.movement.Orientation;
 
 /**
  * Registers input events