+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # UNRELEASED Version 0.4.10
 
+* [[#178](https://github.com/bitbrain/braingdx/issues/178)] add `shake()` method to `GameCamera` to shake the screen
 * [[#180](https://github.com/bitbrain/braingdx/issues/180)] add controller support to `InputManager`
 * cleaning up `ActorTween` API